Extract create value for Color to helper for:
- cmyk
- grayscale
- html
