+++ src/WpfMath.Example/WpfMath.Example.csproj
@@ -8,6 +8,6 @@
         <IsPackable>false</IsPackable>
     </PropertyGroup>
     <ItemGroup>
-        <ProjectReference Include="../WpfMath/WpfMath.csproj"/>
+        <ProjectReference Include="../WpfMath/WpfMath.csproj" />
     </ItemGroup>
 </Project>
